Move the format method which formats a string for a person to be displayed from Messages.java to Person.java
