+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -57,7 +57,7 @@ public class AddCommand extends Command {
         }
 
         model.addPerson(toAdd);
-        return new CommandResult(String.format(MESSAGE_SUCCESS, Messages.format(toAdd)));
+        return new CommandResult(String.format(MESSAGE_SUCCESS, Person.format(toAdd)));
     }
 
     @Override
